requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonify==0.5
MarkupSafe==2.0.1
phonenumbers==8.12.28
Pillow==8.3.1
pycparser==2.20
six==1.16.0
SQLAlchemy==1.4.21
Werkzeug==2.0.1
WTForms==2.3.3
